intellidata/requirements.txt
asgiref==3.2.10
Babel==2.8.0
boto==2.49.0
boto3==1.14.20
botocore==1.17.20
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
Django==3.0.8
django-bootstrap3==12.1.0
django-controlcenter==0.3.0
django-phonenumber-field==4.0.0
django-pkgconf==0.4.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.9
jmespath==0.10.0
misaka==2.1.1
phonenumbers==8.12.6
Pillow==7.2.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.9
